More MSVC warning fixes
git-svn-id: https://svn.osgeo.org/gdal/trunk@31641 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogr_srs_esri.cpp
@@ -2492,7 +2492,7 @@ OGRErr OGRSpatialReference::ImportFromESRIWisconsinWKT( const char* prjName, dou
     {
         if( fabs(centralMeridian - tableWISCRS[i]) <= 0.0000000001 && fabs(latOfOrigin - tableWISCRS[i+1]) <= 0.0000000001) 
         {
-            k = tableWISCRS[i+2];
+            k = (int)tableWISCRS[i+2]; // double to int ? 
             break;
         }
     }
